Rename retry policy files
