+++ sdk/storage/storage-blob/src/policies/StorageRetryPolicy.ts
@@ -15,7 +15,7 @@ import {
   WebResource
 } from "@azure/core-http";
 
-import { StorageRetryOptions } from "../RetryPolicyFactory";
+import { StorageRetryOptions } from "../StorageRetryPolicyFactory";
 import { URLConstants } from "../utils/constants";
 import { delay, setURLHost, setURLParameter } from "../utils/utils.common";
 
@@ -28,8 +28,8 @@ import { delay, setURLHost, setURLParameter } from "../utils/utils.common";
  */
 export function NewRetryPolicyFactory(retryOptions?: StorageRetryOptions): RequestPolicyFactory {
   return {
-    create: (nextPolicy: RequestPolicy, options: RequestPolicyOptions): RetryPolicy => {
-      return new RetryPolicy(nextPolicy, options, retryOptions);
+    create: (nextPolicy: RequestPolicy, options: RequestPolicyOptions): StorageRetryPolicy => {
+      return new StorageRetryPolicy(nextPolicy, options, retryOptions);
     }
   };
 }
@@ -40,7 +40,7 @@ export function NewRetryPolicyFactory(retryOptions?: StorageRetryOptions): Reque
  * @export
  * @enum {number}
  */
-export enum RetryPolicyType {
+export enum StorageRetryPolicyType {
   /**
    * Exponential retry. Retry time delay grows exponentially.
    */
@@ -56,7 +56,7 @@ const DEFAULT_RETRY_OPTIONS: StorageRetryOptions = {
   maxRetryDelayInMs: 120 * 1000,
   maxTries: 4,
   retryDelayInMs: 4 * 1000,
-  retryPolicyType: RetryPolicyType.EXPONENTIAL,
+  retryPolicyType: StorageRetryPolicyType.EXPONENTIAL,
   secondaryHost: "",
   tryTimeoutInMs: undefined // Use server side default timeout strategy
 };
@@ -69,13 +69,13 @@ const RETRY_ABORT_ERROR = new AbortError("The operation was aborted.");
  * @class RetryPolicy
  * @extends {BaseRequestPolicy}
  */
-export class RetryPolicy extends BaseRequestPolicy {
+export class StorageRetryPolicy extends BaseRequestPolicy {
   /**
    * RetryOptions.
    *
    * @private
    * @type {StorageRetryOptions}
-   * @memberof RetryPolicy
+   * @memberof StorageRetryPolicy
    */
   private readonly retryOptions: StorageRetryOptions;
 
@@ -85,7 +85,7 @@ export class RetryPolicy extends BaseRequestPolicy {
    * @param {RequestPolicy} nextPolicy
    * @param {RequestPolicyOptions} options
    * @param {StorageRetryOptions} [retryOptions=DEFAULT_RETRY_OPTIONS]
-   * @memberof RetryPolicy
+   * @memberof StorageRetryPolicy
    */
   constructor(
     nextPolicy: RequestPolicy,
@@ -136,7 +136,7 @@ export class RetryPolicy extends BaseRequestPolicy {
    *
    * @param {WebResource} request
    * @returns {Promise<HttpOperationResponse>}
-   * @memberof RetryPolicy
+   * @memberof StorageRetryPolicy
    */
   public async sendRequest(request: WebResource): Promise<HttpOperationResponse> {
     return this.attemptSendRequest(request, false, 1);
@@ -153,7 +153,7 @@ export class RetryPolicy extends BaseRequestPolicy {
    * @param {number} attempt           How many retries has been attempted to performed, starting from 1, which includes
    *                                   the attempt will be performed by this method call.
    * @returns {Promise<HttpOperationResponse>}
-   * @memberof RetryPolicy
+   * @memberof StorageRetryPolicy
    */
   protected async attemptSendRequest(
     request: WebResource,
@@ -216,7 +216,7 @@ export class RetryPolicy extends BaseRequestPolicy {
    * @param {HttpOperationResponse} [response]
    * @param {RestError} [err]
    * @returns {boolean}
-   * @memberof RetryPolicy
+   * @memberof StorageRetryPolicy
    */
   protected shouldRetry(
     isPrimaryRetry: boolean,
@@ -295,7 +295,7 @@ export class RetryPolicy extends BaseRequestPolicy {
    * @private
    * @param {HttpPipelineLogLevel} _level
    * @param {string} _message
-   * @memberof RetryPolicy
+   * @memberof StorageRetryPolicy
    */
   // tslint:disable-next-line:variable-name
   private logf(_level: HttpPipelineLogLevel, _message: string) {
@@ -309,20 +309,20 @@ export class RetryPolicy extends BaseRequestPolicy {
    * @param {boolean} isPrimaryRetry
    * @param {number} attempt
    * @param {AbortSignalLike} [abortSignal]
-   * @memberof RetryPolicy
+   * @memberof StorageRetryPolicy
    */
   private async delay(isPrimaryRetry: boolean, attempt: number, abortSignal?: AbortSignalLike) {
     let delayTimeInMs: number = 0;
 
     if (isPrimaryRetry) {
       switch (this.retryOptions.retryPolicyType) {
-        case RetryPolicyType.EXPONENTIAL:
+        case StorageRetryPolicyType.EXPONENTIAL:
           delayTimeInMs = Math.min(
             (Math.pow(2, attempt - 1) - 1) * this.retryOptions.retryDelayInMs!,
             this.retryOptions.maxRetryDelayInMs!
           );
           break;
-        case RetryPolicyType.FIXED:
+        case StorageRetryPolicyType.FIXED:
           delayTimeInMs = this.retryOptions.retryDelayInMs!;
           break;
       }
