+++ sdk/storage/storage-file-share/src/index.browser.ts
@@ -21,7 +21,7 @@ export * from "./FileSystemAttributes";
 export * from "./Pipeline";
 export * from "./policies/AnonymousCredentialPolicy";
 export * from "./policies/CredentialPolicy";
-export * from "./RetryPolicyFactory";
+export * from "./StorageRetryPolicyFactory";
 export * from "./TelemetryPolicyFactory";
 export * from "./UniqueRequestIDPolicyFactory";
 export * from "./StorageBrowserPolicyFactory";
