+++ sdk/storage/storage-queue/src/index.ts
@@ -14,7 +14,7 @@ export { SasIPRange } from "./SasIPRange";
 export * from "./Pipeline";
 export * from "./policies/AnonymousCredentialPolicy";
 export * from "./policies/CredentialPolicy";
-export * from "./RetryPolicyFactory";
+export * from "./StorageRetryPolicyFactory";
 export * from "./StorageBrowserPolicyFactory";
 export { Metadata } from "./models";
 export * from "./policies/SharedKeyCredentialPolicy";
